==> Data-Engineering-MLOps-Exam-Assignment/5_online_predictions.ipynb <==
# Online Production

#!pip install -U hopsworks --quiet

import warnings
warnings.filterwarnings("ignore")

import hopsworks

project = hopsworks.login()

# Get the feature store and model serving instances
fs = project.get_feature_store()
ms = project.get_model_serving()


# Fetch Model from Hopsworks Model Registry
deployment = ms.get_deployment(name="beerratingmodeldeployment1")
deployment.start()

## 🚀 Send an Inference Request

# Send inference requests to the deployed model as follows:
feature_view = fs.get_feature_view("user_beer_feature_view", 1)

# Initialize serving
feature_view.init_serving(1)
